fix: split traffic shaper into its own file
